+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # AuxiliaryRails Changelog
 
+## [Unreleased]
+
+### Added
+- Errorable concern
+
 ## v0.3.0
 
 ### Added